tools/datafix/pyproject.toml
[tool.poetry]
name = "datafix"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

google-cloud-ndb = "==2.3.2"
google-cloud-storage = "==2.18.2"
PyYAML = "==6.0.2"

osv = { path = "../../", develop = true }
google-cloud-pubsub = "^2.25.2"

[tool.poetry.dev-dependencies]
yapf = "*"
pylint = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
